Refactor project to add VEP app with a single file upload endpoint for test.
