+++ requirements.txt
@@ -3,3 +3,5 @@ pydantic>=2.4,<2.5
 uvicorn>=0.24,<0.25
 aiohttp>=3.8,<3.9
 vcfpy>=0.13.8
+loguru==0.7.2
+python-multipart==0.0.9
